+++ packages/engine.runtime/specs/modules.d.ts
@@ -1,7 +1,7 @@
-declare type producer = any
-declare const observe: any
-declare const get: any
-declare const update: any
-declare const prop: any
-declare const param: any
-declare const arg: any
+declare type producer = any;
+declare const observe: any;
+declare const get: any;
+declare const update: any;
+declare const prop: any;
+declare const param: any;
+declare const arg: any;